habittracker-ml/pyproject.toml
[project]
name = "habittracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=23.3.0",
    "coremltools>=7.0",
    "flake8>=6.0.0",
    "hydra-core>=1.3.0",
    "matplotlib>=3.7.0",
    "mlflow>=2.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pytest>=7.3.0",
    "scipy>=1.10.0",
    "seaborn>=0.12.0",
    "jupyterlab>=4.5.1",
    "xgboost>=3.1.2",
    "scikit-learn>=1.5.1",
]
